refactor: add vo(value object) package
