+++ src/main/java/lotto/domain/vo/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import java.util.ArrayList;
 import java.util.Collections;
