+++ src/main/java/lotto/veiw/OutputHandler.java
@@ -4,7 +4,7 @@ import static lotto.domain.LottoReference.NOPE;
 
 import java.text.DecimalFormat;
 import java.util.Map;
-import lotto.domain.Lotto;
+import lotto.domain.vo.Lotto;
 import lotto.domain.LottoMachine;
 import lotto.domain.LottoReference;
 import lotto.domain.LottoResult;
